extracted MiniAppSdkConfig to a single file
